execute and burn funcs updates
